UNIXTIME CREATING REMOVED FROM DB FOR NONINDEXING CASE

+++ project/application/source/domain_layer/data/entity/application_user_reset_password_token.rs
@@ -11,8 +11,8 @@ pub struct ApplicationUserResetPasswordToken<'a> {
 }
 
 impl<'a> ApplicationUserResetPasswordToken<'a> {
-    pub const QUANTITY_OF_MINUTES_BEFORE_RESENDING: i16 = 1;
-    pub const QUANTITY_OF_MINUTES_FOR_EXPIRATION: i16 = 10;
+    pub const QUANTITY_OF_MINUTES_BEFORE_RESENDING: i64 = 1;
+    pub const QUANTITY_OF_MINUTES_FOR_EXPIRATION: i64 = 10;
     pub const WRONG_ENTER_TRIES_QUANTITY_LIMIT: i16 = 3;
 
     pub fn new(
